[Build] Use general Spark version helper for source cache
